# pkg/trace/traceutil: Improve perf for large inputs and fix TruncateUTF8