<!--
* New contributors are highly encouraged to read our
  [CONTRIBUTING](/CONTRIBUTING.md) documentation.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's change.

-->
### What does this PR do?
Fixes the behavior of TruncateUTF8 so output never is larger than the byte limit. Also improves performance for large inputs.
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/datadog-agent/pkg/trace/traceutil
                  │   old.out    │               new.out               │
                  │    sec/op    │   sec/op     vs base                │
TruncateString-10   112.60n ± 1%   10.30n ± 0%  -90.85% (p=0.000 n=10)
```
<!--
* A brief description of the change being made with this pull request.
* If the description here cannot be expressed in a succinct form, consider
  opening multiple pull requests instead of a single one.
-->

### Motivation
While going through trace-agent code I happened across this function which seemed like a possible candidate for performance improvement. When I went to try and improve the performance I noticed that the function was actually wrong. For example: "télé" is 6 bytes long in utf-8, with a limit of 5, the correct output should be "tél" at 4 bytes long. 
<!--
* What inspired you to submit this pull request?
* Link any related GitHub issues or PRs here.
-->

### Additional Notes

<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the proposed
  version was chosen.
-->

### Possible Drawbacks / Trade-offs
It's possible someone is relying on this "off" nature of Truncation, however they shouldn't be and the function is documented as returning a string that is SMALLER than `limit`. Additionally this new code is a bit harder to read / understand, I think this downside is sufficiently offset by the new fuzz test which should catch any possible mistakes made in future re-writes/refactors etc.
<!--
* What are the possible side-effects or negative impacts of the code change?
-->

### Describe how to test/QA your changes
I fixed the unit tests as well as added a fuzz test to ensure all the invariants hold. It ran successfully for 3 minutes on my machine not finding new code paths and not finding any failures. Therefore I don't believe this needs separate QA.
<!--
* Write here in detail or link to detailed instructions on how this change can
  be tested/QAd/validated, including any environment setup.
-->

### Reviewer's Checklist
<!--
* Authors can use this list as a reference to ensure that there are no problems
  during the review but the signing off is to be done by the reviewer(s).

Note: Adding GitHub labels is only possible for contributors with write access.
-->

- [ ] If known, an appropriate milestone has been selected; otherwise the `Triage` milestone is set.
- [ ] Use the `major_change` label if your change either has a major impact on the code base, is impacting multiple teams or is changing important well-established internals of the Agent. This label will be use during QA to make sure each team pay extra attention to the changed behavior. For any customer facing change use a releasenote.
- [ ] A [release note](https://github.com/DataDog/datadog-agent/blob/main/docs/dev/contributing.md#reno) has been added or the `changelog/no-changelog` label has been applied.
- [ ] Changed code has automated tests for its functionality.
- [ ] Adequate QA/testing plan information is provided if the `qa/skip-qa` label is not applied.
- [ ] At least one `team/..` label has been applied, indicating the team(s) that should QA this change.
- [ ] If applicable, docs team has been notified or [an issue has been opened on the documentation repo](https://github.com/DataDog/documentation/issues/new).
- [ ] If applicable, the `need-change/operator` and `need-change/helm` labels have been applied.
- [ ] If applicable, the `k8s/<min-version>` label, indicating the lowest Kubernetes version compatible with this feature.
- [ ] If applicable, the [config template](https://github.com/DataDog/datadog-agent/blob/main/pkg/config/config_template.yaml) has been updated.